userStories/requirements.txt
PyYAML
nltk
